Fixed paths to the directories, moved utils functions out of "backend" part
