+++ posts/clever-lazy-loading-for-javascript.md
@@ -1,5 +1,5 @@
 ---
-title: this is ittle
+title: Title for our article
 slug: clever-lazy-loading-for-javascript
 meta: Maybe you’ve heard of it, maybe you haven’t. Zig is a new programming language that seems to be growing in popularity.
 ---
